Prefix fdef:17a0:ffb1:300 on gateway and routers
We used to have a different prefix on both since dnsmasq would kill our default
route otherwise. Now we use a mini-radvd on the routers so this is not an issue
anymore. Now, having different prefixes would often lead to failing
ipv6 now since the clients may select the ipv6 from the router as the source IP
when talking to global ipv6 addresses (as it is not deprecated anymore) but the
gateway then does not know where to route answers to.
The drawback of this solution is that we now use the same prefix as bielefeld.
If we ever were to do any peering, we can not use these addresses.